drop troublesome pep517.meta usage
pep517.meta.load will use pep517.meta.build, which tries to provision a
environment with pip. This is causing lots of issues in downstreams,
namely Arch Linux.
This patch replaces the pep517.meta usage with pypa/build, which can do
the same job without provisioning a isolated environment.

The drawback is that the PEP 517 dependencies will now have to be
present, which I feel is very reasonable.
Given a environment with all the required packages, we should not be
adding a dependency on the network to be able to run this plugin.

Signed-off-by: Filipe Laíns <lains@riseup.net>